+++ database/migrations/2019_03_19_145549_crear_tabla_libro_prestamo.php
@@ -14,10 +14,10 @@ class CrearTablaLibroPrestamo extends Migration
     public function up()
     {
         Schema::create('libro_prestamo', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('usuario_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('usuario_id');
             $table->foreign('usuario_id', 'fk_libroprestamo_usuario')->references('id')->on('usuario')->onDelete('restrict')->onUpdate('restrict');
-            $table->unsignedInteger('libro_id');
+            $table->unsignedBigInteger('libro_id');
             $table->foreign('libro_id', 'fk_libroprestamo_libro')->references('id')->on('libro')->onDelete('restrict')->onUpdate('restrict');
             $table->date('fecha_prestamo');
             $table->string('prestado_a', 100);
